server/requirements.txt
aniso8601==9.0.1
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
PyMySQL==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1
